Moved the client event handlers into a separate file
