+++ pain/src/main/java/net/g3n0/smtpersonamod/block/ModBlocks.java
@@ -1,21 +1,15 @@
 package net.g3n0.smtpersonamod.block;
 
 import net.g3n0.smtpersonamod.SMTPersonaMod;
-import net.g3n0.smtpersonamod.item.ModItems;
 import net.minecraft.block.AbstractBlock;
 import net.minecraft.block.Block;
-import net.minecraft.block.Blocks;
 import net.minecraft.block.OreBlock;
 import net.minecraft.block.material.Material;
-import net.minecraft.item.BlockItem;
-import net.minecraft.item.Item;
 import net.minecraftforge.common.ToolType;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.RegistryObject;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
-
-import java.util.function.Supplier;
 
 public class ModBlocks {
