+++ pain/src/main/java/net/g3n0/smtpersonamod/item/ModItems.java
@@ -2,10 +2,8 @@ package net.g3n0.smtpersonamod.item;
 
 import net.g3n0.smtpersonamod.SMTPersonaMod;
 import net.minecraft.inventory.EquipmentSlotType;
-import net.minecraft.item.ArmorItem;
 import net.minecraft.item.ArmorMaterial;
 import net.minecraft.item.Item;
-import net.minecraft.item.ItemGroup;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.RegistryObject;
 import net.minecraftforge.registries.DeferredRegister;
